+++ apps/website/vite.config.ts
@@ -14,7 +14,6 @@ export default defineConfig({
   resolve: {
     alias: {
       "@docs": path.resolve(directory, "..", "..", "docs"),
-      "@core-docs": path.resolve(directory, "..", "..", "packages", "core", "docs"),
       "@root": path.resolve(directory, "..", ".."),
     },
   },